+++ blake2b/blake2b.bench.ts
@@ -1,6 +1,6 @@
 import * as refImpl from "https://esm.sh/@noble/hashes@1.1.2/blake2b"
-import { Blake2b } from "../blake2b.ts"
-import { benchHasher } from "./bench_util.ts"
+import { benchHasher } from "../bench_util.ts"
+import { Blake2b } from "./blake2b.ts"
 
 benchHasher({
   name: "blake2b",
