+++ common/hasher.ts
