+++ deno.jsonc
@@ -32,5 +32,6 @@
     "bench": "deno task build && deno bench -A --no-check=remote --unstable",
     "star": "deno task run _tasks/star.ts && deno cache --no-check=remote target/star.ts; deno task build",
     "moderate": "deno task run https://deno.land/x/moderate@0.0.5/mod.ts && dprint fmt"
-  }
+  },
+  "lock": false
 }
