--- hashers/mod.ts
@@ -1,5 +1,0 @@
-// moderate --exclude *.wasm.ts
-
-export * from "./blake2b.ts"
-export * from "./common.ts"
-export * from "./xxhash.ts"
