+++ xxhash/xxhash.bench.ts
@@ -1,6 +1,6 @@
 import * as refImpl from "https://esm.sh/@polkadot/util-crypto@10.1.6/xxhash/index.js"
-import { Xxhash } from "../xxhash.ts"
-import { benchHasher } from "./bench_util.ts"
+import { benchHasher } from "../common/bench_util.ts"
+import { Xxhash } from "./xxhash.ts"
 
 for (const size of [512, 64, 128, 256, 384, 512] as const) {
   benchHasher({
